+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.5.2 (unreleased)
 
+- Added `--fail-fast` option
+- Improved error message
 - Fixed typo in error message
 
 ## 0.5.1 (2020-03-26)